+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Filled fields for core DAO entity
+- Added spam field to proposal info
 
 ## [0.4.6] - 2025-03-18