+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.4.7] - 2025-03-18
+
 ### Added
 - Filled fields for core DAO entity
 - Added spam field to proposal info